+++ html/get-snacks.html
@@ -6,6 +6,7 @@
 </head>
 <body>
 <h1>Mad og drikke</h1>
+<button type="submit" data-id="">Tilføj</button>
 <table id="snackTable">
     <thead>
     <tr>
@@ -22,6 +23,6 @@
 </table>
 <div id="errorMessage"></div>
 
-<script type="module" src="../js/getSnacks.js"></script>
+<script type="module" src="../js/get-snacks.js"></script>
 </body>
 </html>
